bharti intern
Titanic_Survival_Classification
This is a classic project that involves predicting whether a passenger on the Titanic survived or not based on various features

Project goals:

To build a simple machine learning model that can predict the survival of Titanic passengers.

Project steps: 1.Collect the Titanic passenger data. 2.Clean and prepare the data for machine learning. 3.Train a machine learning model on the data. 4.Evaluate the accuracy of the model. 5.Analyze the features of the model to identify the factors that are most important for predicting survival.

Project resources:

1.The Titanic passenger data is available from the Kaggle website. 2.The Python programming language will be used for machine learning. 3.The seaborn,pandas,matplotlib this python library are used.

By using simple machine learning techinques to predict the survival of Titanic passengers, we can gain insights into the factors that influence survival in emergency situations. This information can be used to improve safety regulations and to help people make better decisions in the future. This project is developed as a part of data science internship by bharat Intern.